Use a guard statement pattern to flatten code
Standard if-else statements have generally been kept at the final branch
which I think provides the best balance of readability.

Because this is all touching the same code, HTTP return codes have also
been corrected. HTTP 401 should be returned if a token is somehow
invalid because the authentication itself is being rejected whereas,
for example, if an account is suspended an HTTP 403 should be returned
because authentication was successful but accessing further features is
not permitted.